User Input to manipulate an image/box
Trying to make it so WASD can move an image and have collisions with other images on the screen.

Soon, trying to make it so there is a 3D camera perspective that will be controlled with this same WASD + mouse cursor position.

Will keep the mouse cursor locked in the center of the window until they tab out or exit.