Collisions and Movement
Added a basic loop to move a square (or any shape) around.

Added a renderer to create a circle by connecting many points with each (x, y) being plotted with the radians' ratio at each segment.

Added collisions on the X axis. Planning on adding one on the Y axis.

Working on implementing a way to layer screens onto a Window object. Screens will either work as a regular ArrayList or as a stack (to go backwards and forwards)